+++ modules/apim-adapter/src/main/java/com/axway/apim/api/definition/OAS3xSpecification.java
@@ -1,7 +1,11 @@
 package com.axway.apim.api.definition;
 
+import java.net.MalformedURLException;
+import java.net.URL;
+
 import com.axway.apim.lib.CoreParameters;
 import com.axway.apim.lib.errorHandling.AppException;
+import com.axway.apim.lib.errorHandling.ErrorCode;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
@@ -28,14 +32,17 @@ public class OAS3xSpecification extends APISpecification {
 		if(!CoreParameters.getInstance().isReplaceHostInSwagger()) return;
 		try {
 			if(backendBasepath!=null) {
-				 ObjectNode newServer = this.mapper.createObjectNode();
-				 newServer.put("url", backendBasepath);
-				 if(openAPI.has("servers")) {
+				new URL(backendBasepath); // Parse it to make sure it is valid
+				ObjectNode newServer = this.mapper.createObjectNode();
+				newServer.put("url", backendBasepath);
+				if(openAPI.has("servers")) {
 					((ArrayNode) openAPI.get("servers")).removeAll();
-				 }
-				 ((ObjectNode)openAPI).set("servers", mapper.createArrayNode().add(newServer));
+				}
+				((ObjectNode)openAPI).set("servers", mapper.createArrayNode().add(newServer));
 				this.apiSpecificationContent = this.mapper.writeValueAsBytes(openAPI);
 			}
+		} catch (MalformedURLException e) {
+			throw new AppException("The backendBasepath: '"+backendBasepath+"' is invalid.", ErrorCode.CANT_READ_CONFIG_FILE, e);
 		} catch (Exception e) {
 			LOG.error("Cannot replace host in provided Swagger-File. Continue with given host.", e);
 		}