+++ modules/apim-adapter/src/main/java/com/axway/apim/api/definition/Swagger1xSpecification.java
@@ -1,9 +1,11 @@
 package com.axway.apim.api.definition;
 
+import java.net.MalformedURLException;
 import java.net.URL;
 
 import com.axway.apim.lib.CoreParameters;
 import com.axway.apim.lib.errorHandling.AppException;
+import com.axway.apim.lib.errorHandling.ErrorCode;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -44,6 +46,8 @@ public class Swagger1xSpecification extends APISpecification {
 				}
 				this.apiSpecificationContent = this.mapper.writeValueAsBytes(swagger);
 			}
+		} catch (MalformedURLException e) {
+			throw new AppException("The backendBasepath: '"+backendBasepath+"' is invalid.", ErrorCode.CANT_READ_CONFIG_FILE, e);
 		} catch (Exception e) {
 			LOG.error("Cannot replace host in provided Swagger-File. Continue with given host.", e);
 		}