+++ modules/apis/src/test/java/com/axway/apim/api/definition/APISpecificationOpenAPI3xTest.java
@@ -57,6 +57,14 @@ public class APISpecificationOpenAPI3xTest {
 		Assert.assertEquals( ((ArrayNode) swagger.get("servers")).get(0).get("url").asText(), "https://myhost.customer.com:8767/api/v1/myAPI");
 	}
 	
+	@Test(expectedExceptions = AppException.class, expectedExceptionsMessageRegExp = "The backendBasepath: 'An-Invalid-URL' is invalid.")
+	public void testInvalidBackendBasepath() throws AppException, IOException {
+
+		byte[] content = getSwaggerContent(testPackage + "/openapi30.json");
+		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
+		apiDefinition.configureBasepath("An-Invalid-URL");
+	}
+	
 	
 	private byte[] getSwaggerContent(String swaggerFile) throws AppException {
 		try {