+++ modules/apis/src/test/java/com/axway/apim/api/definition/APISpecificationSwagger1xTest.java
@@ -80,7 +80,13 @@ public class APISpecificationSwagger1xTest {
 		Assert.assertEquals(swagger.get("basePath").asText(), "https://petstore.swagger.io:443/myapi");
 	}
 	
-	
+	@Test(expectedExceptions = AppException.class, expectedExceptionsMessageRegExp = "The backendBasepath: 'An-Invalid-URL' is invalid.")
+	public void testInvalidBackendBasepath() throws AppException, IOException {
+
+		byte[] content = getSwaggerContent(testPackage + "/swagger11.json");
+		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
+		apiDefinition.configureBasepath("An-Invalid-URL");
+	}
 	
 	
 	private byte[] getSwaggerContent(String swaggerFile) throws AppException {