+++ modules/apis/src/test/java/com/axway/apim/api/definition/APISpecificationSwagger2xTest.java
@@ -125,6 +125,14 @@ public class APISpecificationSwagger2xTest {
 		Assert.assertEquals(swagger.get("schemes").size(), 2);
 	}
 	
+	@Test(expectedExceptions = AppException.class, expectedExceptionsMessageRegExp = "The backendBasepath: 'An-Invalid-URL' is invalid.")
+	public void testInvalidBackendBasepath() throws AppException, IOException {
+
+		byte[] content = getSwaggerContent("/api_definition_1/petstore.json");
+		APISpecification apiDefinition = APISpecificationFactory.getAPISpecification(content, "teststore.json", "Test-API");
+		apiDefinition.configureBasepath("An-Invalid-URL");
+	}
+	
 	
 	private byte[] getSwaggerContent(String swaggerFile) throws AppException {
 		try {